Added calculate diff step before stagepipelines

+++ tests/drone/copy-files-for-upload.sh
@@ -2,6 +2,10 @@
 # Input parameters
 # $1 web directory
 
-ls -la /filesForUpload
-cp -a "$1"/tests/acceptance/filesForUpload/. /filesForUpload
-ls -la /filesForUpload
+if test -f runUnitTestsOnly
+then echo 'Bye!'
+else
+	ls -la /filesForUpload
+	cp -a "$1"/tests/acceptance/filesForUpload/. /filesForUpload
+	ls -la /filesForUpload
+fi
